NationStates

Documentation

data Context

Keeps track of rate limits and TLS connections.

You should create a single Context at the start of your program, then share it between multiple threads and requests.

newContext

Arguments

:: String
User agent

-> IO Context

Create a new Context.

newContext'

Arguments

:: String
User agent

-> Bool
Use HTTPS

-> IO Context

Create a new Context, with extra options.

Core types

module NationStates.Types