NationStates.RateLimit

Description

Simple rate limiting combinator.

Documentation

data RateLimit

newRateLimit

Arguments

:: Rational
Delay, in seconds

-> IO RateLimit

Create a new rate limiter with the specified delay.

The rate limiter is thread-safe, and can be shared between threads.

rateLimit :: RateLimit -> IO a -> IO a

Run the given action, pausing as necessary to keep under the rate limit.

setDelay :: Rational -> RateLimit -> RateLimit

Create a new rate limiter with the same lock but a different delay.