NationStates.Types

Description

Data structures used by NationStates.

Documentation

data WACategory

Nation category.

This datum summarizes a nation's personal, economic, and political freedoms.

Constructors

Anarchy

AuthoritarianDemocracy

BenevolentDictatorship

CapitalistParadise

Capitalizt

CivilRightsLovefest

CompulsoryConsumeristState

ConservativeDemocracy

CorporateBordello

CorporatePoliceState

CorruptDictatorship

DemocraticSocialists

FatherKnowsBestState FatherOrMother
This category has two variations: "Father Knows Best State" and "Mother Knows Best State".

FreeMarketParadise

InoffensiveCentristDemocracy

IronFistConsumerists

IronFistSocialists

LeftLeaningCollegeState

LeftWingUtopia

LiberalDemocraticSocialists

LibertarianPoliceState

MoralisticDemocracy

NewYorkTimesDemocracy

PsychoticDictatorship

RightWingUtopia

ScandinavianLiberalParadise

TyrannyByMajority

data FatherOrMother

Differentiates between a Father or Mother Knows Best State.

Constructors

Father

Mother

readWACategory :: String -> Maybe WACategory

showWACategory :: WACategory -> String

type WAVote = Maybe Bool

A vote in the General Assembly or Security Council.

Just True and Just False mean "for" and "against" respectively, while Nothing means that the nation did not vote at all.

readWAVote :: String -> Maybe WAVote

readWAVote' :: String -> Maybe (Maybe WAVote)

showWAVote :: WAVote -> String

readWAStatus :: String -> Maybe Bool

showWAStatus :: Bool -> String